+++ src/domain/kakao/kakao.module.ts
@@ -2,12 +2,14 @@ import {forwardRef, Module} from '@nestjs/common';
 import { KakaoController } from './kakao.controller';
 import { ChatModule } from '../chat/chat.module';
 import {HttpModule} from "@nestjs/axios";
-import {UserSessionService} from "./user-session.service";
+import {UserSessionService} from "../user/user-session.service";
 import {ArticleQueueModule} from "./queue/article-queue.module";
 import {BullModule} from "@nestjs/bull";
+import {UserModule} from "../user/user.module";
+import {PostModule} from "../post/post.module";
 
 @Module({
-    imports: [ChatModule, HttpModule,
+    imports: [ChatModule, HttpModule, UserModule, PostModule,
         forwardRef(() => ArticleQueueModule),
         BullModule.registerQueue({ name: 'articleQueue' }),],
     controllers: [KakaoController],
